rename events to activiy
