+++ client/src/components/activity/view.vue
